rtlwifi_new:Rework cmd_send_packet
The incorrect placement of a lock meant that the TX ring queue was not
protected. On one system, this led to freezes when running rtl8192ee.
When checking the other drivers to see which ones are affected, it was
discovered that several of them used identical code. Accordingly, this
routine was moved into core.c where all the drivers could share a single
copy.
